+++ src/content/index.ts
@@ -1,5 +1,7 @@
 import { init as initHideExplore } from './hide-explore';
 import { init as initHideReels } from './hide-reels';
+import { init as initFilterFeed } from './filter-feed';
 
 initHideExplore();
 initHideReels();
+initFilterFeed();